requirements.txt
hupper
pymongo
pyyaml
python-telegram-bot
aiohttp
langchain
langchain-ollama
langchain-openrouter